cambios charlados en grupo.-

+++ TpIngSoftII/Models/Entities/Empleado.cs
@@ -11,7 +11,7 @@ namespace TpIngSoftII.Models.Entities
         public int ID { get; set; }
         public string Nombre { get; set; }
         public string Apellido { get; set; }
-        public int Dni { get; set; }
+        public long Dni { get; set; }
         public DateTime FechaIngreso { get; set; }
         public string Usuario { get; set; }
         public string Clave { get; set; }
